Move web server code to main crate
